Stats de defesa convertido para float normalizado

+++ Time-3/Assets/Scripts/Personagens/CharacterBase.cs
@@ -9,7 +9,7 @@ public class CharacterBase : ScriptableObject
 {
     [Header("Atributos do jogador")]
     public int baseHp;
-    public int baseDefense;
+    public float baseDefense;
     public float baseSpeed;
 
     [Tooltip("Dano do ataque basico")]
@@ -19,7 +19,7 @@ public class CharacterBase : ScriptableObject
     public int baseSkillDmg;
 
     public SkillBase basicAttack;
-    
+
     public SkillBase[] Skills;
     [Header("Sprites")]
     public Sprite head;
@@ -28,3 +28,4 @@ public class CharacterBase : ScriptableObject
     public RuntimeAnimatorController animatorController;
 
 }
+
